Drop jump predictions
Although there are branch likely instruction the whole mips instruction
set is designed around static jump predictions. The usage of delay slot
ensures that we don't have to handle jump predictions on cpu it self at
all. Adding it on top of that makes little to no sense although that was
the original plan.