Implement little bit more of scheme and fix connection angle
This commit adds few more bits to scheme but mainly it chnages how
connectors specify angles. Originally it was in radians but we was
mapping that trough mathematical operations directly to sizes. But that
was problematic because of floating point inacuracy and we sometimes
founded intersection where there should be one. So this commit gets rid
of this at all and instead allows just some fixes axes to be used
instead of arbitrary angles.