Disable program follow
This should be easy enought to program but it requires some deeper
changes to memoryview it self (to allows direct address to item
mapping). Because of that I am disabling it for now and I will return to
this feature in future. For now I am focusing on other more important
aspects of qtmips (to make it feature complete not nice to use).