Fix nested calls of setCurrentIndex which caused breakage.
ProgramTableView::focus_address() calls QAbstractItemView::setCurrentIndex().
verticalScrollBar() value is updated as result of current row change.
This emits signal valueChanged which is connected to ProgramTableView::adjust_scroll_pos().
It checks if the limit of range covered by actual model and row to address offset
is reached. If the top or bottom 1/8 of range is reached then model
needs to be adjusted to cover continuation area. Model shift requires
update of the current row to stay on the same address even that row 0
address offset is changed.

This model shifting is required because range of scroll is only signed
integer and QTableView is even more limited in row count to work
reliably.

Signed-off-by: Pavel Pisa <pisa@cmp.felk.cvut.cz>