GUI: invert the process of component installation
Instead of searching the whole tree for each component type, the tree is searched only once, finding all nodes with `data-component` attribute and dispatching the installer code on the attribute value. This is much faster as DOM is traversed only once, it removes a hard-to-comprehend layer of abstraction and allows for non-uniform installer code (i.e. passing different arguments to each component) to adapt for component controllers diverging from regular, one updatable text, structure.