svgscene: qt6 - qtringref was removed
QStringRef was removed in Qt6 and it is encouraged to use QStringView
instead, hovewer QStringView is not present in Qt5.9.5. Therefore in
this version it is polyfilled using QString. This will result in reduced
performance, but I don't care. If the user uses such old system they
should be happy to have any kind of support. This polyfill also
complicates API as QString does not have some useful methods like
`chopped`.
TLDR: I cannot wait to drop Ubuntu 18 support.