CLI: Moved main function of CLI into QT's main loop
Before this commit, it could occur that QCoreApplication::exit() is called before the call to before QCoreApplication::exec(). As a result, the call to QCoreApplication::exit() gets ignored and QtRVSim will never terminate. A simple way to reproduce the bug is to call QtRVSim on an empty assembly file: `qtrvsim_cli --asm empty.S`.

This commit fixes the issue by moving initialization code (including the argument parsing and assembler invocation) into the Qt main loop. For that, code from main.cpp is moved into a new class CLIMain. Some previously local variables of the main function are now members of CLIMain.